refactor: changes folder structure to header
