+++ src/components/Header/DropdownMenu.js
@@ -1,0 +1,6 @@
+import React from "react"
+import * as S from "../../styles/style.js"
+
+export function DropdownMenu(props) {
+  return <S.DropdownMenu>{props.children}</S.DropdownMenu>
+}
