+++ src/components/Header/NavItem.js
@@ -1,0 +1,5 @@
+import React from "react"
+
+export function NavItem(props) {
+  return <li>{props.children}</li>
+}
