+++ src/components/Header/NavItemHidden.js
@@ -1,0 +1,7 @@
+import React, { useContext } from "react"
+import { MenuContext } from "../../contexts/MenuContext.js"
+
+export function NavItemHidden(props) {
+  const { menuIsOpen } = useContext(MenuContext)
+  return <>{menuIsOpen && props.children}</>
+}
