+++ src/components/Header/Navbar.js
@@ -1,0 +1,9 @@
+import React from "react"
+
+export function Navbar(props) {
+  return (
+    <nav>
+      <ul> {props.children}</ul>
+    </nav>
+  )
+}
