+++ src/layouts/DefaultPage.js
@@ -1,7 +1,7 @@
 import React from "react"
 import { Outlet } from "react-router-dom"
 
-import Header from "../components/Header.js"
+import Header from "../components/Header/Header.js"
 
 export default function DefaultPage() {
   return (
